==> Analyses/Scripts/0. Data import.Rmd <==
---
title: "0. Data import"
author: "Lauke"
date: "12/01/2022"
output: html_document
---

#Convert large PISA data files to smaller file format & subset

```{r Libraries}
library(haven)
library(tidyverse)
library(readxl)
```

## Cognitive data
```{r Read in - warning: big file takes long}
memory.limit(size = 1048576)
COGdat <- read_sav("Data/SPSS-COG/CY07_MSU_STU_COG.sav")
```

```{r Select booklet}
book1 <- COGdat %>% filter(BOOKID == 1)
```

```{r Read in item information}
items <- read_excel("Data/PISA2018 TechRep_Final-AnnexA.xlsx", sheet = 3) #get the excel in R
math_items <- items %>%
  pull(1)  #pull the first column
math_items <- math_items[7:length(math_items)] #first 6 entries are not items, so exclude those
```

```{r Construct booklet 1 math file}
#select only the math items in the data, plus some information
pisa18_math <- book1 %>%
  mutate_at(vars(all_of(math_items)), as_factor) %>%
  select(c(CNT, CNTSCHID, CNTSTUID, BOOKID, all_of(math_items))) #leaves 86 cols of which 82 items (checks out)

#delete the empty columns, i.e. items that were not administered in booklet 1
book1_math <- pisa18_math[colSums(!is.na(pisa18_math)) > 0] #leaves 27 cols of which 23 items
```

```{r Save data file}
save(book1_math, file = "Workspaces/book1_math.Rdata")
```

## Student questionnaire data
```{r Read in - warning: big file takes long}
QQQdat_stu <- read_sav("Data/SPSS-QQQ/CY07_MSU_STU_QQQ.sav")
```

```{r Subset QQQ data to match selected booklet}
QQQbook1_math <- QQQdat_stu %>% filter(CNTSTUID %in% book1_math$CNTSTUID)
```

```{r Save QQQ data file}
save(QQQbook1_math, file = "Workspaces/QQQbook1_math.Rdata")
```

## Country level data

```{r Create CNT code to CNT label object}
cnt_label <- read_excel("Data/PISA2018_CODEBOOK.xlsx", sheet = 2)
cnt_label <- cnt_label[88:169, 7:8]
colnames(cnt_label) <- c("CNT", "Country")
```

```{r Age of first selection}
sel_age <- read_excel("Data/PISA 2018 Results (Volume I)_AnnexB3.XLSX", sheet = 11) #get the excel in R

#create column names
sel_age[12,1] <- "Country"
colnames(sel_age) <- sel_age[12,]

#add variable containing OECD status: 1 = member, 0 = partner
sel_age$OECD_status <- rep(c(1,0), c(53,48))

#drop empty cells
sel_age <- sel_age[c(16:52, 55:96), -3]

#match CNT code with label
sel_age <- inner_join(cnt_label, sel_age, by = "Country")

#shorten column name
colnames(sel_age)[4:5] <- c("nr_schooltypes", "selection_age")

#save vars as numeric
sel_age$nr_schooltypes <- as.numeric(sel_age$nr_schooltypes)  
sel_age$selection_age <- as.numeric(sel_age$selection_age)
#NA's enforced were the non-OECD "m" entries. nothing lost. 
```

```{r Save selection age file}
save(sel_age, file = "Workspaces/sel_age.Rdata")
```

```{r Remove redudant objects from workspace}
rm(COGdat, book1, math_items, pisa18_math, QQQdat_stu, cnt_label)
```

```{r Print session info}
sessionInfo()
```
