==> Analyses/Scripts/1. Data pre-processing.Rmd <==
---
title: "1. Data pre-processing"
author: "Lauke"
date: "22/03/2022"
output: html_document
---

```{r Libraries}
library(haven)
library(dplyr)
library(tidyr)
library(kableExtra)
```

```{r Load in data}
load("~/GitHub/Master_Thesis/Analyses/Workspaces/QQQbook1_math.Rdata")
load("~/GitHub/Master_Thesis/Analyses/Workspaces/sel_age.Rdata")
```

```{r Prep data step 1}
#Merge selection age info with QQQ data
step1_dat <- inner_join(sel_age[,c(1,2,4,5)], QQQbook1_math, by = "CNT") %>% 
  filter(OECD == 1)  %>% #select only OECD members
  dplyr::select(CNTSTUID, selection_age, nr_schooltypes, AGE, ESCS, GRADE, ISCEDL, ST004D01T, ST022Q01TA) %>% #select used vars only
  rename(Gender = ST004D01T, Language = ST022Q01TA) %>% 
  drop_na()
    
table(step1_dat$selection_age) #dichotomies at 11, 12, 14 and 15 

step1_dat$group11 <- ifelse(step1_dat$selection_age <= 11, 0, 1) #if selection age < or = 11, group 0, >11 group 1.
step1_dat$group12 <- ifelse(step1_dat$selection_age <= 12, 0, 1) #if selection age < or = 12, group 0, >12 group 1.

step1_dat$group14 <- ifelse(step1_dat$selection_age <= 14, 0, 1) #if selection age < or = 14, group 0, >14 group 1.
step1_dat$group15 <- ifelse(step1_dat$selection_age <= 15, 0, 1) #if selection age < or = 15, group 0, >15 group 1.

all.equal(step1_dat$group12, ifelse(step1_dat$selection_age <= 13, 0, 1)) #NB: 
```
```{r Save data step 1}
save(step1_dat, file = "Workspaces/step1_dat.Rdata")
```


```{r Latex table observations per age - Table 1 in manuscript}
results <- as.data.frame(table(step1_dat$selection_age)) %>% mutate(Percentages = (Freq/(sum(Freq)))*100)
freqtable <- results %>%
  bind_rows(summarise_all(., ~if(is.numeric(.)) sum(.) else "Total"))

results

kbl(
  freqtable,
  booktabs = T,
  col.names = c("Age at first selection", "Frequency", "Percentage of total"),
  label = "freq_sel_age",
  format = "latex", 
  digits = 0, 
  caption = "Frequencies of age at first selection in OECD countries")
```

```{r Latex table selection age + groups - Table B1 in Appendix B manuscript}
kbl(sel_age %>% filter(OECD_status == 1) %>% 
        dplyr::select(Country, selection_age) %>% 
        mutate(group11 = ifelse(selection_age <= 11, 0, 1),
               group12 = ifelse(selection_age <= 12, 0, 1),
               group14 = ifelse(selection_age <= 14, 0, 1),
               group15 = ifelse(selection_age <= 15, 0, 1)), 
    format = "latex",
    longtable = T,
    booktabs = T,
    col.names = c("Country", "Age at first selection", "Age 11", "Age 12", "Age 14", "Age 15"), 
    caption = "Age at first selection of each OECD country", 
    label = "sel_age") %>% 
  add_header_above(c("", "", "Treatment assignment by age dichotomy" = 4)) %>% 
  footnote(general = "Source: Annex 3B of the PISA 2018 Results: Volume 1") %>% 
  kable_styling(latex_options = "repeat_header")
```


```{r Prep data step 2: add ability estimates to dataset}
#obtain 10 plausible values on the MATH items per pupil from the original dataset
PVs <- QQQbook1_math %>% dplyr::select(CNTSTUID, starts_with("PV") & ends_with("MATH")) #select PVs for all students who took maths items
PVs$ability <- rowMeans(PVs[,2:11]) #compute ability estimate per pupil
```

```{r Save PV data}
save(PVs, file = "Workspaces/PVs.Rdata")
```


```{r Remove redudant objects from workspace}
rm(QQQbook1_math, sel_age, results)
```

```{r Print session info}
sessionInfo()
```
