==> Analyses/Scripts/Step 1 - Propenstiy score estimation.Rmd <==
---
title: "Step 1 - Propensity score estimation"
author: "Lauke"
date: "28/01/2022"
output: html_document
---

```{r Set global seed}
set.seed(123)
```


```{r load in data}
load("~/GitHub/Master_Thesis/Analyses/Workspaces/step1_dat.Rdata")
```


```{r Libraries}
library(dplyr)
library(haven)
library(knitr)
library(kableExtra)
library(MCMCpack) #needed for PSM procedure MCMClogit function
library(mcmcplots) #multiple traceplots in 1 frame
library(MatchIt)
library(ggplot2)
library(ggpubr) #to make grid plot
library(colorBlindness) #to check for colour blindness
```

Step 1: obtain a propensity score distribution for each pupil. 

# Balance check 'before' USING MATCHIT

```{r Balance check group = 11}
balance_pre11 <- summary(matchit(
      group11 ~ as.factor(Gender) + AGE + GRADE + ISCEDL + ESCS + as.factor(Language),
      method = NULL, #null method for 'pre-matching' balance
      estimand = "ATE", 
      data = step1_dat, 
    ))
```

```{r Balance check group = 12}
balance_pre12 <- summary(matchit(
      group12 ~ as.factor(Gender) + AGE + GRADE + ISCEDL + ESCS + as.factor(Language),
      method = NULL, #null method for 'pre-matching' balance
      estimand = "ATE", 
      data = step1_dat, 
    ))
```

```{r Balance check group = 14}
balance_pre14 <- summary(matchit(
      group14 ~ as.factor(Gender) + AGE + GRADE + ISCEDL + ESCS + as.factor(Language),
      method = NULL, #null method for 'pre-matching' balance
      estimand = "ATE", 
      data = step1_dat, 
    ))
```

```{r Balance check group = 15}
balance_pre15 <- summary(matchit(
      group15 ~ as.factor(Gender) + AGE + GRADE + ISCEDL + ESCS + as.factor(Language),
      method = NULL, #null method for 'pre-matching' balance
      estimand = "ATE", 
      data = step1_dat, 
    ))
```


```{r Table on dataset age 12 - Table 2 in manuscript}
kbl(as.data.frame(balance_pre12$sum.all)[-1,1:4], digits = 2, format = "latex", col.names = c("Mean Treated", "Mean Control", "SMD", "VR"), caption = "Balance check on all data before matching - treatment assignment age > 12", booktabs = T, label = "balance_pre_12") 
```


```{r Table on all datasets (horizontal) - not used in manuscript}
dat <- cbind(as.data.frame(balance_pre11$sum.all)[-1,1:4],
             as.data.frame(balance_pre12$sum.all)[-1,1:4],
             as.data.frame(balance_pre14$sum.all)[-1,1:4],
             as.data.frame(balance_pre15$sum.all)[-1,1:4])

kbl(dat, booktabs = T, format = "latex", caption = "Balance check on all data before matching all ages", 
    col.names = rep(c("Mean Treated", "Mean Control", "SMD", "VR"), 4), label = "balance_pre_all") %>%
add_header_above(c(" ", "Age 11" = 4, "Age 12" = 4, "Age 14" = 4, "Age 15" = 4)) %>%
kable_styling(latex_options = c("scale_down"))
```

```{r Table on all datasets (vertical) - Table B2 in Appendix B manuscript}
vertdat <- rbind(as.data.frame(balance_pre11$sum.all)[-1,1:4],
             as.data.frame(balance_pre12$sum.all)[-1,1:4],
             as.data.frame(balance_pre14$sum.all)[-1,1:4],
             as.data.frame(balance_pre15$sum.all)[-1,1:4])

kbl(vertdat, booktabs = T, longtable = T, format = "latex", digits = 2, caption = "Balance check on all data before matching all ages", label = "balance_pre_all") %>%
kable_styling(position = "center", latex_options = c("repeat_header")) %>% 
  pack_rows("Age 11", 1, 8, latex_gap_space = "2em") %>% 
    pack_rows("Age 12", 9, 16, latex_gap_space = "2em") %>% 
    pack_rows("Age 14", 17, 24, latex_gap_space = "2em") %>% 
    pack_rows("Age 15", 25, 32, latex_gap_space = "2em")

```


# PROPSENSITY SCORE ESTIMATION

## AGE 11
```{r Estimate posterior coefficients for alpha & betas - AGE 11}
set.seed(123)
posterior_11 <- MCMClogit(group11 ~ as.factor(Gender) + AGE + GRADE + ISCEDL + ESCS + as.factor(Language), mcmc = 2000, burnin = 1000, thin = 2, data = step1_dat)

summary(posterior_11)

#assess convergence
traplot(posterior_11)
autocorr.plot(posterior_11)

#colourblindness check
cvdPlot(traplot(posterior_11))
```

```{r Estimate propensity scores for each pupil based on posterior samples of alpha and beta - AGE 11}
#create matrix containing all independent variables + intercept (1)
X_matrix_11 <- step1_dat %>% 
  mutate(constant = 1) %>% #add column to be multiplied by intercept
  dplyr::select(constant, Gender, AGE, GRADE, ISCEDL, ESCS, Language) %>% #select intercept + Xvars
  t() %>% #transpose to fit posterior object (col = pupils, row = vars)
  as.matrix() #store as matrix

#multiply the obtained posterior estimates with IV's to obtain propensity scores
prop_scores_11 <- exp(posterior_11 %*% X_matrix_11) / #logistic regression, see manuscript equation 3.
  (1 + exp(posterior_11 %*% X_matrix_11)) #col = pupils, row = n.iter prop scores

colnames(prop_scores_11) <- step1_dat$CNTSTUID #give cols student ID names

hist(colMeans(prop_scores_11))
mean(colMeans(prop_scores_11))
table(step1_dat$group11)
```

## AGE 12
```{r Estimate posterior coefficients for alpha & betas - AGE 12}
set.seed(123)
posterior_12 <- MCMClogit(group12 ~ as.factor(Gender) + AGE + GRADE + ISCEDL + ESCS + as.factor(Language), mcmc = 2000, burnin = 1000, thin = 2, data = step1_dat)

summary(posterior_12)

#assess convergence
traplot(posterior_12)
autocorr.plot(posterior_12)
```

```{r Estimate propensity scores for each pupil based on posterior samples of alpha and beta - AGE 12}
#create matrix containing all independent variables + intercept (1)
X_matrix_12 <- step1_dat %>% 
  mutate(constant = 1) %>% #add column to be multiplied by intercept
  dplyr::select(constant, Gender, AGE, GRADE, ISCEDL, ESCS, Language) %>% #select intercept + Xvars
  t() %>% #transpose to fit posterior object (col = pupils, row = vars)
  as.matrix() #store as matrix

#multiply the obtained posterior estimates with IV's to obtain propensity scores
prop_scores_12 <- exp(posterior_12 %*% X_matrix_12) / #logistic regression, see manuscript equation 3.
  (1 + exp(posterior_12 %*% X_matrix_12)) #col = pupils, row = n.iter prop scores

colnames(prop_scores_12) <- step1_dat$CNTSTUID #give cols student ID names

hist(colMeans(prop_scores_12)) 
mean(colMeans(prop_scores_12)) 
table(step1_dat$group12)
```

## AGE 14
```{r Estimate posterior coefficients for alpha & betas - AGE 14}
set.seed(123)
posterior_14 <- MCMClogit(group14 ~ as.factor(Gender) + AGE + GRADE + ISCEDL + ESCS + as.factor(Language), mcmc = 2000, burnin = 1000, thin = 2, data = step1_dat)

summary(posterior_14)

#assess convergence
traplot(posterior_14)
autocorr.plot(posterior_14)
```

```{r Estimate propensity scores for each pupil based on posterior samples of alpha and beta - AGE 14}
#create matrix containing all independent variables + intercept (1)
X_matrix_14 <- step1_dat %>% 
  mutate(constant = 1) %>% #add column to be multiplied by intercept
  dplyr::select(constant, Gender, AGE, GRADE, ISCEDL, ESCS, Language) %>% #select intercept + Xvars
  t() %>% #transpose to fit posterior object (col = pupils, row = vars)
  as.matrix() #store as matrix

#multiply the obtained posterior estimates with IV's to obtain propensity scores
prop_scores_14 <- exp(posterior_14 %*% X_matrix_14) / #logistic regression, see manuscript equation 3.
  (1 + exp(posterior_14 %*% X_matrix_14)) #col = pupils, row = n.iter prop scores

colnames(prop_scores_14) <- step1_dat$CNTSTUID #give cols student ID names

hist(colMeans(prop_scores_14)) 
mean(colMeans(prop_scores_14)) 
table(step1_dat$group14)
```

## AGE 15
```{r Estimate posterior coefficients for alpha & betas - AGE 15}
set.seed(123)
posterior_15 <- MCMClogit(group15 ~ as.factor(Gender) + AGE + GRADE + ISCEDL + ESCS + as.factor(Language), mcmc = 2000, burnin = 1000, thin = 2, data = step1_dat)

summary(posterior_15)

#assess convergence
traplot(posterior_15)
autocorr.plot(posterior_15)
```


```{r Estimate propensity scores for each pupil based on posterior samples of alpha and beta - AGE 15}
#create matrix containing all independent variables + intercept (1)
X_matrix_15 <- step1_dat %>% 
  mutate(constant = 1) %>% #add column to be multiplied by intercept
  dplyr::select(constant, Gender, AGE, GRADE, ISCEDL, ESCS, Language) %>% #select intercept + Xvars
  t() %>% #transpose to fit posterior object (col = pupils, row = vars)
  as.matrix() #store as matrix

#multiply the obtained posterior estimates with IV's to obtain propensity scores
prop_scores_15 <- exp(posterior_15 %*% X_matrix_15) / #logistic regression, see manuscript equation 3.
  (1 + exp(posterior_15 %*% X_matrix_15)) #col = pupils, row = n.iter prop scores

colnames(prop_scores_15) <- step1_dat$CNTSTUID #give cols student ID names

hist(colMeans(prop_scores_15)) 
mean(colMeans(prop_scores_15)) 
table(step1_dat$group15)
```

```{r Save propensity score files}
save(prop_scores_11, file = "Workspaces/prop_scores_11.Rdata")
save(prop_scores_12, file = "Workspaces/prop_scores_12.Rdata")
save(prop_scores_14, file = "Workspaces/prop_scores_14.Rdata")
save(prop_scores_15, file = "Workspaces/prop_scores_15.Rdata")
```


# VISUALISATION of overlap propensity score distributions

```{r Propensity scores by treatment assignment}
prop_scores_by_group11 <- data.frame(mean_propscores = colMeans(prop_scores_11), #mean ps per student
                         CNTSTUID = as.numeric(colnames(prop_scores_11))) %>%  #extract student ID
  left_join(step1_dat %>% dplyr::select(CNTSTUID, group11), by = "CNTSTUID") #join ps and student ID with group membership

prop_scores_by_group12 <- data.frame(mean_propscores = colMeans(prop_scores_12),
                         CNTSTUID = as.numeric(colnames(prop_scores_12))) %>% 
  left_join(step1_dat %>% dplyr::select(CNTSTUID, group12), by = "CNTSTUID") 

prop_scores_by_group14 <- data.frame(mean_propscores = colMeans(prop_scores_14),
                         CNTSTUID = as.numeric(colnames(prop_scores_14))) %>% 
  left_join(step1_dat %>% dplyr::select(CNTSTUID, group14), by = "CNTSTUID") 

prop_scores_by_group15 <- data.frame(mean_propscores = colMeans(prop_scores_15),
                         CNTSTUID = as.numeric(colnames(prop_scores_15))) %>% 
  left_join(step1_dat %>% dplyr::select(CNTSTUID, group15), by = "CNTSTUID") 
```

```{r SD of prop scores for group = 12}
mean_sd_prop_scores_group12 <- data.frame(sd_propscores = prop_scores_12 %>% as.data.frame() %>% apply(., 2, sd)) %>% colMeans()
```


```{r Plots for all dichotomies}
safecolours <- c("#117733", "#44AA99")

plot11 <- ggplot() + 
  geom_histogram(data = filter(prop_scores_by_group11, group11 == 1), 
                 bins = 50, aes(x = mean_propscores), 
                 fill = "#117733") + 
  geom_histogram(data = filter(prop_scores_by_group11, group11 == 0), 
                 bins = 50, aes(x = mean_propscores, y = -..count..),
                 fill = "#44AA99") +
  geom_hline(yintercept = 0) +
  ggtitle("Propensity score distribution Age AFS > 11") +
  theme(plot.title = element_text(hjust = 0.5)) +
  annotate(geom = "label", x = 0, y = 50, label = "Treat", 
           fill = "#117733", color = "white", hjust = 0) +
  annotate(geom = "label", x = 0, y = -50, label = "Control", 
           fill = "#44AA99", color = "white", hjust = 0) +
  scale_y_continuous(label = abs) +
  coord_cartesian(xlim = c(0, 1), ylim = c(-300, 950)) +
  labs(x = "Propensity score", y = "Count")

plot12 <- ggplot() + 
  geom_histogram(data = filter(prop_scores_by_group12, group12 == 1), 
                 bins = 50, aes(x = mean_propscores), 
                 fill = "#117733") + 
  geom_histogram(data = filter(prop_scores_by_group12, group12 == 0), 
                 bins = 50, aes(x = mean_propscores, y = -..count..),
                 fill = "#44AA99") +
  geom_hline(yintercept = 0) +
  ggtitle("Propensity score distribution Age AFS > 12") +
  theme(plot.title = element_text(hjust = 0.5)) +
  annotate(geom = "label", x = 0, y = 50, label = "Treat", 
           fill = "#117733", color = "white", hjust = 0) +
  annotate(geom = "label", x = 0, y = -50, label = "Control", 
           fill = "#44AA99", color = "white", hjust = 0) +
  scale_y_continuous(label = abs) +
  coord_cartesian(xlim = c(0, 1), ylim = c(-300, 950)) +
  labs(x = "Propensity score", y = "Count")


plot14 <- ggplot() + 
  geom_histogram(data = filter(prop_scores_by_group14, group14 == 1), 
                 bins = 50, aes(x = mean_propscores), 
                 fill = "#117733") + 
  geom_histogram(data = filter(prop_scores_by_group14, group14 == 0), 
                 bins = 50, aes(x = mean_propscores, y = -..count..),
                 fill = "#44AA99") +
  geom_hline(yintercept = 0) +
  ggtitle("Propensity score distribution Age AFS > 14") +
  theme(plot.title = element_text(hjust = 0.5)) +
  annotate(geom = "label", x = 0, y = 50, label = "Treat", 
           fill = "#117733", color = "white", hjust = 0) +
  annotate(geom = "label", x = 0, y = -50, label = "Control", 
           fill = "#44AA99", color = "white", hjust = 0) +
  scale_y_continuous(label = abs) +
  coord_cartesian(xlim = c(0, 1), ylim = c(-300, 950)) +
  labs(x = "Propensity score", y = "Count")

plot15 <- ggplot() + 
  geom_histogram(data = filter(prop_scores_by_group15, group15 == 1), 
                 bins = 50, aes(x = mean_propscores), 
                 fill = "#117733") + 
  geom_histogram(data = filter(prop_scores_by_group15, group15 == 0), 
                 bins = 50, aes(x = mean_propscores, y = -..count..),
                 fill = "#44AA99") +
  geom_hline(yintercept = 0) +
  ggtitle("Propensity score distribution Age AFS > 15") +
  theme(plot.title = element_text(hjust = 0.5)) +
  annotate(geom = "label", x = 0, y = 50, label = "Treat", 
           fill = "#117733", color = "white", hjust = 0) +
  annotate(geom = "label", x = 0, y = -50, label = "Control", 
           fill = "#44AA99", color = "white", hjust = 0) +
  scale_y_continuous(label = abs) +
  coord_cartesian(xlim = c(0, 1), ylim = c(-300, 950)) +
  labs(x = "Propensity score", y = "Count")

```

```{r Colour blindness check}
cvdPlot(plot12)
```

```{r Plot all together in 1 figure - Figure 1 in manuscript}
plotgrid <- ggarrange(plot11, plot12, plot14, plot15,
          labels = "AUTO",
          ncol = 2, nrow = 2)

save(plotgrid, file = "Figures/plotgrid.Rdata")
```


```{r Remove redudant objects from workspace}
rm(list=setdiff(ls(), c("step1_dat", "prop_scores_11", "prop_scores_12", "prop_scores_14", "prop_scores_15")))
```

```{r Print session info}
sessionInfo()
```
